% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{F_to_K}
\alias{F_to_K}
\title{Fahrenheit conversion}
\usage{
F_to_K(F_temp)
}
\arguments{
\item{F_temp}{The temperature in degrees Fahrenheit}
}
\value{
The temperature in degrees Kelvin
}
\description{
Convert degrees Fahrenheit temperatures to degrees Kelvin
}
\examples{
temp1 <- F_to_K(50);
temp2 <- F_to_K( c(50, 63, 23) );
}
