% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{K_to_F}
\alias{K_to_F}
\title{Kelvin conversion}
\usage{
K_to_F(K_temp)
}
\arguments{
\item{C_temp}{The temperature in degrees Kelvin}
}
\value{
The temperature in degrees Fahrenheit
}
\description{
Convert degrees Kelvin temperatures to degrees Fahrenheit
}
\examples{
temp1 <- K_to_F(22);
temp2 <- K_to_F( c(0, 12, 23) );
}
